Generalized Linear Models

OptimizationProblems.GeneralizedLinearModel — Type

GeneralizedLinearModel{R, F, G<:GLMFamily} <: OptimizationProblem

Data for specifying the negative log-likelihood objective function for a generalized linear model.

Fields

name::String, the name of the problem

counters::Dict{Symbol, Counter}, a dictionary of symbols that identify a counter

num_param::Int64, dimension of the optimization parameter.

num_obs::Int64, the total number of observations.

resp::R, the responses of the data of type R, which will depend on the type of model being considered.

feat::F, the features or explanatory variables of type F, which will depend on the way features are stored.

family::G, specifies the GLM family being considered

Constructors

LogisticRegression

For details, see the docstrings for each function listed.

OptimizationProblems.GLMFamily — Type

GLMFamily

An abstract type specifying the partition function for a GLM.

Logistic Regression Problem

OptimizationProblems.LogisticRegression — Function

LogisticRegression(::Type{T}; num_param::Int64, num_obs::Int64, 
    name::String="Logistic Regression") where T <: Real

Constructs a Logistic Regression Problem where the given number of parameters is num_param and the number of observations is num_obs. The feature matrix is a matrix of type T. It has a column of ones followed by num_param-1 columns of independent random normal vectors with mean zero and variance 1/num_param. The response vector is a randomly generated BitVector corresponding to a logistic regression model. Returns a GeneralizedLinearModel{BitVector, Matrix{T}, Bernoulli}.

LogisticRegression(;resp::BitVector, feat::Matrix{T},
    name::String="Logistic Regression") where T <: Real

Constructs a Logistic Regression Problem with the user-supplied response vector, resp, and feature matrix, feat. Returns a GeneralizedLinearModel{BitVector, Matrix{T}, Bernoulli}.

OptimizationProblems.Bernoulli — Type

Bernoulli <: GLMFamily

A structure specifying a Bernoulli response GLM.