Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

OptimizationProblems.meta

OptimizationProblems.PureJuMP.AMPGO02

OptimizationProblems.PureJuMP.AMPGO03

OptimizationProblems.PureJuMP.AMPGO04

OptimizationProblems.PureJuMP.AMPGO05

OptimizationProblems.PureJuMP.AMPGO06

OptimizationProblems.PureJuMP.AMPGO07

OptimizationProblems.PureJuMP.AMPGO08

OptimizationProblems.PureJuMP.AMPGO09

OptimizationProblems.PureJuMP.AMPGO10

OptimizationProblems.PureJuMP.AMPGO11

OptimizationProblems.PureJuMP.AMPGO12

OptimizationProblems.PureJuMP.AMPGO13

OptimizationProblems.PureJuMP.AMPGO14

OptimizationProblems.PureJuMP.AMPGO15

OptimizationProblems.PureJuMP.AMPGO18

OptimizationProblems.PureJuMP.AMPGO20

OptimizationProblems.PureJuMP.AMPGO21

OptimizationProblems.PureJuMP.AMPGO22

OptimizationProblems.PureJuMP.Dus2_1

OptimizationProblems.PureJuMP.Dus2_3

OptimizationProblems.PureJuMP.Dus2_9

OptimizationProblems.PureJuMP.Duscube

OptimizationProblems.PureJuMP.Shpak1

OptimizationProblems.PureJuMP.Shpak2

OptimizationProblems.PureJuMP.Shpak3

OptimizationProblems.PureJuMP.Shpak4

OptimizationProblems.PureJuMP.Shpak5

OptimizationProblems.PureJuMP.Shpak6

OptimizationProblems.PureJuMP.arglina

OptimizationProblems.PureJuMP.arglinb

OptimizationProblems.PureJuMP.arglinc

OptimizationProblems.PureJuMP.arwhead

OptimizationProblems.PureJuMP.bdqrtic

OptimizationProblems.PureJuMP.beale

OptimizationProblems.PureJuMP.brownbs

OptimizationProblems.PureJuMP.brownden

OptimizationProblems.PureJuMP.broydn7d

OptimizationProblems.PureJuMP.brybnd

OptimizationProblems.PureJuMP.bt1

OptimizationProblems.PureJuMP.chainwoo

OptimizationProblems.PureJuMP.cliff

OptimizationProblems.PureJuMP.clnlbeam

OptimizationProblems.PureJuMP.clplatea

OptimizationProblems.PureJuMP.clplateb

OptimizationProblems.PureJuMP.clplatec

OptimizationProblems.PureJuMP.cosine

OptimizationProblems.PureJuMP.cragglvy

OptimizationProblems.PureJuMP.curly

OptimizationProblems.PureJuMP.curly10

OptimizationProblems.PureJuMP.curly20

OptimizationProblems.PureJuMP.curly30

OptimizationProblems.PureJuMP.dixmaane

OptimizationProblems.PureJuMP.dixmaanf

OptimizationProblems.PureJuMP.dixmaang

OptimizationProblems.PureJuMP.dixmaanh

OptimizationProblems.PureJuMP.dixmaani

OptimizationProblems.PureJuMP.dixmaanj

OptimizationProblems.PureJuMP.dixmaank

OptimizationProblems.PureJuMP.dixmaanl

OptimizationProblems.PureJuMP.dixmaanm

OptimizationProblems.PureJuMP.dixmaann

OptimizationProblems.PureJuMP.dixmaano

OptimizationProblems.PureJuMP.dixmaanp

OptimizationProblems.PureJuMP.dixon3dq

OptimizationProblems.PureJuMP.dqdrtic

OptimizationProblems.PureJuMP.dqrtic

OptimizationProblems.PureJuMP.edensch

OptimizationProblems.PureJuMP.eg2

OptimizationProblems.PureJuMP.engval1

OptimizationProblems.PureJuMP.genrose

OptimizationProblems.PureJuMP.genrose_nash

OptimizationProblems.PureJuMP.hs1

OptimizationProblems.PureJuMP.hs10

OptimizationProblems.PureJuMP.hs100

OptimizationProblems.PureJuMP.hs101

OptimizationProblems.PureJuMP.hs102

OptimizationProblems.PureJuMP.hs103

OptimizationProblems.PureJuMP.hs104

OptimizationProblems.PureJuMP.hs105

OptimizationProblems.PureJuMP.hs106

OptimizationProblems.PureJuMP.hs107

OptimizationProblems.PureJuMP.hs108

OptimizationProblems.PureJuMP.hs109

OptimizationProblems.PureJuMP.hs11

OptimizationProblems.PureJuMP.hs110

OptimizationProblems.PureJuMP.hs111

OptimizationProblems.PureJuMP.hs112

OptimizationProblems.PureJuMP.hs113

OptimizationProblems.PureJuMP.hs114

OptimizationProblems.PureJuMP.hs116

OptimizationProblems.PureJuMP.hs117

OptimizationProblems.PureJuMP.hs118

OptimizationProblems.PureJuMP.hs119

OptimizationProblems.PureJuMP.hs12

OptimizationProblems.PureJuMP.hs13

OptimizationProblems.PureJuMP.hs14

OptimizationProblems.PureJuMP.hs15

OptimizationProblems.PureJuMP.hs16

OptimizationProblems.PureJuMP.hs17

OptimizationProblems.PureJuMP.hs18

OptimizationProblems.PureJuMP.hs19

OptimizationProblems.PureJuMP.hs2

OptimizationProblems.PureJuMP.hs20

OptimizationProblems.PureJuMP.hs21

OptimizationProblems.PureJuMP.hs22

OptimizationProblems.PureJuMP.hs23

OptimizationProblems.PureJuMP.hs24

OptimizationProblems.PureJuMP.hs25

OptimizationProblems.PureJuMP.hs26

OptimizationProblems.PureJuMP.hs27

OptimizationProblems.PureJuMP.hs28

OptimizationProblems.PureJuMP.hs29

OptimizationProblems.PureJuMP.hs3

OptimizationProblems.PureJuMP.hs30

OptimizationProblems.PureJuMP.hs31

OptimizationProblems.PureJuMP.hs32

OptimizationProblems.PureJuMP.hs33

OptimizationProblems.PureJuMP.hs34

OptimizationProblems.PureJuMP.hs35

OptimizationProblems.PureJuMP.hs36

OptimizationProblems.PureJuMP.hs37

OptimizationProblems.PureJuMP.hs38

OptimizationProblems.PureJuMP.hs39

OptimizationProblems.PureJuMP.hs4

OptimizationProblems.PureJuMP.hs40

OptimizationProblems.PureJuMP.hs41

OptimizationProblems.PureJuMP.hs42

OptimizationProblems.PureJuMP.hs43

OptimizationProblems.PureJuMP.hs44

OptimizationProblems.PureJuMP.hs45

OptimizationProblems.PureJuMP.hs46

OptimizationProblems.PureJuMP.hs47

OptimizationProblems.PureJuMP.hs48

OptimizationProblems.PureJuMP.hs49

OptimizationProblems.PureJuMP.hs5

OptimizationProblems.PureJuMP.hs50

OptimizationProblems.PureJuMP.hs51

OptimizationProblems.PureJuMP.hs52

OptimizationProblems.PureJuMP.hs53

OptimizationProblems.PureJuMP.hs54

OptimizationProblems.PureJuMP.hs55

OptimizationProblems.PureJuMP.hs56

OptimizationProblems.PureJuMP.hs57

OptimizationProblems.PureJuMP.hs59

OptimizationProblems.PureJuMP.hs6

OptimizationProblems.PureJuMP.hs60

OptimizationProblems.PureJuMP.hs61

OptimizationProblems.PureJuMP.hs62

OptimizationProblems.PureJuMP.hs63

OptimizationProblems.PureJuMP.hs64

OptimizationProblems.PureJuMP.hs65

OptimizationProblems.PureJuMP.hs66

OptimizationProblems.PureJuMP.hs7

OptimizationProblems.PureJuMP.hs70

OptimizationProblems.PureJuMP.hs71

OptimizationProblems.PureJuMP.hs72

OptimizationProblems.PureJuMP.hs73

OptimizationProblems.PureJuMP.hs74

OptimizationProblems.PureJuMP.hs75

OptimizationProblems.PureJuMP.hs76

OptimizationProblems.PureJuMP.hs77

OptimizationProblems.PureJuMP.hs78

OptimizationProblems.PureJuMP.hs79

OptimizationProblems.PureJuMP.hs8

OptimizationProblems.PureJuMP.hs80

OptimizationProblems.PureJuMP.hs81

OptimizationProblems.PureJuMP.hs83

OptimizationProblems.PureJuMP.hs84

OptimizationProblems.PureJuMP.hs86

OptimizationProblems.PureJuMP.hs9

OptimizationProblems.PureJuMP.hs93

OptimizationProblems.PureJuMP.hs95

OptimizationProblems.PureJuMP.hs96

OptimizationProblems.PureJuMP.hs97

OptimizationProblems.PureJuMP.hs98

OptimizationProblems.PureJuMP.meyer3

OptimizationProblems.PureJuMP.nasty

OptimizationProblems.PureJuMP.palmer1c

OptimizationProblems.PureJuMP.palmer1d

OptimizationProblems.PureJuMP.palmer2c

OptimizationProblems.PureJuMP.palmer3c

OptimizationProblems.PureJuMP.palmer4c

OptimizationProblems.PureJuMP.palmer5c

OptimizationProblems.PureJuMP.palmer5d

OptimizationProblems.PureJuMP.palmer6c

OptimizationProblems.PureJuMP.palmer7c

OptimizationProblems.PureJuMP.palmer8c

OptimizationProblems.PureJuMP.penalty2

OptimizationProblems.PureJuMP.penalty3

OptimizationProblems.PureJuMP.powellsg

OptimizationProblems.PureJuMP.power

OptimizationProblems.PureJuMP.quartc

OptimizationProblems.PureJuMP.rosenbrock

OptimizationProblems.PureJuMP.sbrybnd

OptimizationProblems.PureJuMP.schmvett

OptimizationProblems.PureJuMP.scosine

OptimizationProblems.PureJuMP.sinquad

OptimizationProblems.PureJuMP.sparsine

OptimizationProblems.PureJuMP.sparsqur

OptimizationProblems.PureJuMP.srosenbr

OptimizationProblems.PureJuMP.tointgss

OptimizationProblems.PureJuMP.tquartic

OptimizationProblems.PureJuMP.tridia

OptimizationProblems.PureJuMP.vardim

OptimizationProblems.PureJuMP.woods

OptimizationProblems.PureJuMP.zangwil3

​

OptimizationProblems.meta — Constant

OptimizationProblems.meta A composite type that represents the main features of the optimization problem. optimize obj(x) subject to lvar ≤ x ≤ uvar lcon ≤ cons(x) ≤ ucon –- The following keys are valid:

nvar::Int: number of variables

variable_nvar::Bool: true if we can modify the number of variables

ncon::Int: number of general constraints

variable_ncon::Bool: true if we can modify the number of constraints

minimize::Bool: true if optimize == minimize

name::String: problem name

has_equalities_only::Bool: true if the problem has constraints, and all are equality constraints (doesn't include bounds)

has_inequalities_only::Bool: true if the problem has constraints, and all are inequality constraints (doesn't include bounds)

has_bounds::Bool: true if the problem has bound constraints

has_fixed_variables::Bool: true if it has fixed variables

objtype::Symbol: type of objective, in [:none, :constant, :linear, :quadratic, :sumofsquares, :other]

contype::Symbol: type of constraint, in [:unconstrained, :linear, :quadratic, :general]

best_known_lower_bound::Real: lower bound on the global optimal value (default: -Inf for minimization problem, f(x0) for maximization problem if x0 is feasible, -Inf otherwise)

best_known_upper_bound::Real: upper bound on the global optimal value (default: Inf for maximization problem, f(x0) for minimization problem if x0 is feasible, Inf otherwise)

is_feasible::Union{Bool, Missing}: true if problem is feasible

origin::Symbol: origin of the problem, in [:academic, :modelling, :real, :unknown]

OptimizationProblems.PureJuMP.AMPGO02 — Method

Univariate multimodal minimization problem AMPGO02

OptimizationProblems.PureJuMP.AMPGO03 — Method

Univariate multimodal minimization problem AMPGO03

OptimizationProblems.PureJuMP.AMPGO04 — Method

Univariate multimodal minimization problem AMPGO04

OptimizationProblems.PureJuMP.AMPGO05 — Method

Univariate multimodal minimization problem AMPGO05

OptimizationProblems.PureJuMP.AMPGO06 — Method

Univariate multimodal minimization problem AMPGO06

OptimizationProblems.PureJuMP.AMPGO07 — Method

Univariate multimodal minimization problem AMPGO07

OptimizationProblems.PureJuMP.AMPGO08 — Method

Univariate multimodal minimization problem AMPGO08

OptimizationProblems.PureJuMP.AMPGO09 — Method

Univariate multimodal minimization problem AMPGO09

OptimizationProblems.PureJuMP.AMPGO10 — Method

Univariate multimodal minimization problem AMPGO10

OptimizationProblems.PureJuMP.AMPGO11 — Method

Univariate multimodal minimization problem AMPGO11

OptimizationProblems.PureJuMP.AMPGO12 — Method

Univariate multimodal minimization problem AMPGO12

OptimizationProblems.PureJuMP.AMPGO13 — Method

Univariate multimodal minimization problem AMPGO13

OptimizationProblems.PureJuMP.AMPGO14 — Method

Univariate multimodal minimization problem AMPGO14

OptimizationProblems.PureJuMP.AMPGO15 — Method

Univariate multimodal minimization problem AMPGO15

OptimizationProblems.PureJuMP.AMPGO18 — Method

Univariate multimodal minimization problem AMPGO18

OptimizationProblems.PureJuMP.AMPGO20 — Method

Univariate multimodal minimization problem AMPGO20

OptimizationProblems.PureJuMP.AMPGO21 — Method

Univariate multimodal minimization problem AMPGO21

OptimizationProblems.PureJuMP.AMPGO22 — Method

Univariate multimodal minimization problem AMPGO22

OptimizationProblems.PureJuMP.Dus2_1 — Method

Univariate unimodal minimization problem Dus2_1

OptimizationProblems.PureJuMP.Dus2_3 — Method

Univariate unimodal minimization problem Dus2_3

OptimizationProblems.PureJuMP.Dus2_9 — Method

Univariate multimodal minimization problem Dus2_9

OptimizationProblems.PureJuMP.Duscube — Method

Univariate multimodal minimization problem Duscube

OptimizationProblems.PureJuMP.Shpak1 — Method

Univariate multimodal minimization problem Shpak1

OptimizationProblems.PureJuMP.Shpak2 — Method

Univariate multimodal minimization problem Shpak2

OptimizationProblems.PureJuMP.Shpak3 — Method

Univariate multimodal minimization problem Shpak3

OptimizationProblems.PureJuMP.Shpak4 — Method

Univariate multimodal minimization problem Shpak4

OptimizationProblems.PureJuMP.Shpak5 — Method

Univariate multimodal minimization problem Shpak5

OptimizationProblems.PureJuMP.Shpak6 — Method

Univariate multimodal minimization problem Shpak6

OptimizationProblems.PureJuMP.arglina — Method

Linear function with n parameters and m observations - full rank

OptimizationProblems.PureJuMP.arglinb — Method

Linear function with n parameters and m observations - rank 1

OptimizationProblems.PureJuMP.arglinc — Method

Linear function with n parameters and m observations - rank 1, zero columns and rows

OptimizationProblems.PureJuMP.arwhead — Method

Arrow head model in size n

OptimizationProblems.PureJuMP.bdqrtic — Method

Banded quartic model in size n

OptimizationProblems.PureJuMP.beale — Method

Beale Model

OptimizationProblems.PureJuMP.brownbs — Method

Brownbs Model

OptimizationProblems.PureJuMP.brownden — Method

Brown and Dennis function

OptimizationProblems.PureJuMP.broydn7d — Method

Broyden 7-diagonal model in size n

OptimizationProblems.PureJuMP.brybnd — Method

Broyden banded model in size n

OptimizationProblems.PureJuMP.bt1 — Method

BT1 Model

OptimizationProblems.PureJuMP.chainwoo — Method

The chained Woods function in size n, a variant on the Woods function

OptimizationProblems.PureJuMP.cliff — Method

The 'cliff problem' in 2 variables

OptimizationProblems.PureJuMP.clnlbeam — Method

The clnlbeam problem in size n

OptimizationProblems.PureJuMP.clplatea — Method

The clamped plate problem (Strang, Nocedal, Dax).

OptimizationProblems.PureJuMP.clplateb — Method

The clamped plate problem (Strang, Nocedal, Dax).

OptimizationProblems.PureJuMP.clplatec — Method

The clamped plate problem (Strang, Nocedal, Dax).

OptimizationProblems.PureJuMP.cosine — Method

The cosine function in size n

OptimizationProblems.PureJuMP.cragglvy — Method

The extented Cragg and Levy function in size n

OptimizationProblems.PureJuMP.curly — Method

Curly function in size n with semi-bandwidth b

OptimizationProblems.PureJuMP.curly10 — Method

Curly function in size n with semi-bandwidth 10

OptimizationProblems.PureJuMP.curly20 — Method

Curly function in size n with semi-bandwidth 20

OptimizationProblems.PureJuMP.curly30 — Method

Curly function in size n with semi-bandwidth 30

OptimizationProblems.PureJuMP.dixmaane — Method

Dixon-Maany function in size n (version E by default)

OptimizationProblems.PureJuMP.dixmaanf — Method

Dixon-Maany function in size n (version F)

OptimizationProblems.PureJuMP.dixmaang — Method

Dixon-Maany function in size n (version G)

OptimizationProblems.PureJuMP.dixmaanh — Method

Dixon-Maany function in size n (version H)

OptimizationProblems.PureJuMP.dixmaani — Method

Dixon-Maany function in size n (version I by default)

OptimizationProblems.PureJuMP.dixmaanj — Method

Dixon-Maany function in size n (version J)

OptimizationProblems.PureJuMP.dixmaank — Method

Dixon-Maany function in size n (version K)

OptimizationProblems.PureJuMP.dixmaanl — Method

Dixon-Maany function in size n (version L)

OptimizationProblems.PureJuMP.dixmaanm — Method

Dixon-Maany function in size n (version M by default)

OptimizationProblems.PureJuMP.dixmaann — Method

Dixon-Maany function in size n (version N)

OptimizationProblems.PureJuMP.dixmaano — Method

Dixon-Maany function in size n (version O)

OptimizationProblems.PureJuMP.dixmaanp — Method

Dixon-Maany function in size n (version P)

OptimizationProblems.PureJuMP.dixon3dq — Method

Dixon's tridiagonal quadratic.

OptimizationProblems.PureJuMP.dqdrtic — Method

Diagonal quadratic problem

OptimizationProblems.PureJuMP.dqrtic — Method

Diagonal quartic model in size n

OptimizationProblems.PureJuMP.edensch — Method

Extended Dennis-Schnabel model in size n

OptimizationProblems.PureJuMP.eg2 — Method

model in size n

OptimizationProblems.PureJuMP.engval1 — Method

The Engval1 model in size n

OptimizationProblems.PureJuMP.genrose — Method

Generalized Rosenbrock model in size n

OptimizationProblems.PureJuMP.genrose_nash — Method

Nash's variant of genrose() in size n

OptimizationProblems.PureJuMP.hs1 — Method

HS1 model

OptimizationProblems.PureJuMP.hs10 — Method

HS10 model

OptimizationProblems.PureJuMP.hs100 — Method

HS100 model

OptimizationProblems.PureJuMP.hs101 — Method

HS101 model

OptimizationProblems.PureJuMP.hs102 — Method

HS102 model

OptimizationProblems.PureJuMP.hs103 — Method

HS103 model

OptimizationProblems.PureJuMP.hs104 — Method

HS104 model

OptimizationProblems.PureJuMP.hs105 — Method

HS105 model

OptimizationProblems.PureJuMP.hs106 — Method

HS106 model

OptimizationProblems.PureJuMP.hs107 — Method

HS107 model

OptimizationProblems.PureJuMP.hs108 — Method

HS108 model

OptimizationProblems.PureJuMP.hs109 — Method

HS109 model

OptimizationProblems.PureJuMP.hs11 — Method

HS11 model

OptimizationProblems.PureJuMP.hs110 — Method

HS110 model

OptimizationProblems.PureJuMP.hs111 — Method

HS111 model

OptimizationProblems.PureJuMP.hs112 — Method

HS112 model

OptimizationProblems.PureJuMP.hs113 — Method

HS113 model

OptimizationProblems.PureJuMP.hs114 — Method

HS114 model

OptimizationProblems.PureJuMP.hs116 — Method

HS116 model

OptimizationProblems.PureJuMP.hs117 — Method

HS117 model

OptimizationProblems.PureJuMP.hs118 — Method

HS118 model

OptimizationProblems.PureJuMP.hs119 — Method

HS119 model

OptimizationProblems.PureJuMP.hs12 — Method

HS12 model

OptimizationProblems.PureJuMP.hs13 — Method

HS13 model

OptimizationProblems.PureJuMP.hs14 — Method

HS14 model

OptimizationProblems.PureJuMP.hs15 — Method

HS15 model

OptimizationProblems.PureJuMP.hs16 — Method

HS16 model

OptimizationProblems.PureJuMP.hs17 — Method

HS17 model

OptimizationProblems.PureJuMP.hs18 — Method

HS18 model

OptimizationProblems.PureJuMP.hs19 — Method

HS19 model

OptimizationProblems.PureJuMP.hs2 — Method

HS2 model

OptimizationProblems.PureJuMP.hs20 — Method

HS20 model

OptimizationProblems.PureJuMP.hs21 — Method

HS21 model

OptimizationProblems.PureJuMP.hs22 — Method

HS22 model

OptimizationProblems.PureJuMP.hs23 — Method

HS23 model

OptimizationProblems.PureJuMP.hs24 — Method

HS24 model

OptimizationProblems.PureJuMP.hs25 — Method

HS25 model

OptimizationProblems.PureJuMP.hs26 — Method

HS26 Model

OptimizationProblems.PureJuMP.hs27 — Method

HS27 Model

OptimizationProblems.PureJuMP.hs28 — Method

HS28 model

OptimizationProblems.PureJuMP.hs29 — Method

HS29 model

OptimizationProblems.PureJuMP.hs3 — Method

HS3 model

OptimizationProblems.PureJuMP.hs30 — Method

HS30 model

OptimizationProblems.PureJuMP.hs31 — Method

HS31 model

OptimizationProblems.PureJuMP.hs32 — Method

HS32 model

OptimizationProblems.PureJuMP.hs33 — Method

HS33 model

OptimizationProblems.PureJuMP.hs34 — Method

HS34 model

OptimizationProblems.PureJuMP.hs35 — Method

HS35 model

OptimizationProblems.PureJuMP.hs36 — Method

HS36 model

OptimizationProblems.PureJuMP.hs37 — Method

HS37 model

OptimizationProblems.PureJuMP.hs38 — Method

HS38 model

OptimizationProblems.PureJuMP.hs39 — Method

HS39 model

OptimizationProblems.PureJuMP.hs4 — Method

HS4 model

OptimizationProblems.PureJuMP.hs40 — Method

HS40 model

OptimizationProblems.PureJuMP.hs41 — Method

HS41 model

OptimizationProblems.PureJuMP.hs42 — Method

HS42 model

OptimizationProblems.PureJuMP.hs43 — Method

HS43 model

OptimizationProblems.PureJuMP.hs44 — Method

HS44 model

OptimizationProblems.PureJuMP.hs45 — Method

HS45 model

OptimizationProblems.PureJuMP.hs46 — Method

HS46 model

OptimizationProblems.PureJuMP.hs47 — Method

HS47 model

OptimizationProblems.PureJuMP.hs48 — Method

HS48 model

OptimizationProblems.PureJuMP.hs49 — Method

HS49 model

OptimizationProblems.PureJuMP.hs5 — Method

HS5 model

OptimizationProblems.PureJuMP.hs50 — Method

HS50 model

OptimizationProblems.PureJuMP.hs51 — Method

HS51 model

OptimizationProblems.PureJuMP.hs52 — Method

HS52 model

OptimizationProblems.PureJuMP.hs53 — Method

HS53 model

OptimizationProblems.PureJuMP.hs54 — Method

HS54 model

OptimizationProblems.PureJuMP.hs55 — Method

HS55 model

OptimizationProblems.PureJuMP.hs56 — Method

HS56 model

OptimizationProblems.PureJuMP.hs57 — Method

HS57 model

OptimizationProblems.PureJuMP.hs59 — Method

HS59 model

OptimizationProblems.PureJuMP.hs6 — Method

HS6 model

OptimizationProblems.PureJuMP.hs60 — Method

HS60 model

OptimizationProblems.PureJuMP.hs61 — Method

HS61 model

OptimizationProblems.PureJuMP.hs62 — Method

HS62 model

OptimizationProblems.PureJuMP.hs63 — Method

HS63 model

OptimizationProblems.PureJuMP.hs64 — Method

HS64 model

OptimizationProblems.PureJuMP.hs65 — Method

HS65 model

OptimizationProblems.PureJuMP.hs66 — Method

HS66 model

OptimizationProblems.PureJuMP.hs7 — Method

HS7 model

OptimizationProblems.PureJuMP.hs70 — Method

HS70 model

OptimizationProblems.PureJuMP.hs71 — Method

HS71 model

OptimizationProblems.PureJuMP.hs72 — Method

HS72 model

OptimizationProblems.PureJuMP.hs73 — Method

HS73 model

OptimizationProblems.PureJuMP.hs74 — Method

HS74 model

OptimizationProblems.PureJuMP.hs75 — Method

HS75 model

OptimizationProblems.PureJuMP.hs76 — Method

HS76 model

OptimizationProblems.PureJuMP.hs77 — Method

HS77 model

OptimizationProblems.PureJuMP.hs78 — Method

HS78 model

OptimizationProblems.PureJuMP.hs79 — Method

HS79 model

OptimizationProblems.PureJuMP.hs8 — Method

HS8 model

OptimizationProblems.PureJuMP.hs80 — Method

HS80 model

OptimizationProblems.PureJuMP.hs81 — Method

HS81 model

OptimizationProblems.PureJuMP.hs83 — Method

HS83 model

OptimizationProblems.PureJuMP.hs84 — Method

HS84 model

OptimizationProblems.PureJuMP.hs86 — Method

HS86 model

OptimizationProblems.PureJuMP.hs9 — Method

HS9 Model

OptimizationProblems.PureJuMP.hs93 — Method

HS93 model

OptimizationProblems.PureJuMP.hs95 — Method

HS95 model

OptimizationProblems.PureJuMP.hs96 — Method

HS96 model

OptimizationProblems.PureJuMP.hs97 — Method

HS97 model

OptimizationProblems.PureJuMP.hs98 — Method

HS98 model

OptimizationProblems.PureJuMP.meyer3 — Method

Meyer function

OptimizationProblems.PureJuMP.nasty — Method

Nasty problem

OptimizationProblems.PureJuMP.palmer1c — Method

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.PureJuMP.palmer1d — Method

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.PureJuMP.palmer2c — Method

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.PureJuMP.palmer3c — Method

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.PureJuMP.palmer4c — Method

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.PureJuMP.palmer5c — Method

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.PureJuMP.palmer5d — Method

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.PureJuMP.palmer6c — Method

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.PureJuMP.palmer7c — Method

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.PureJuMP.palmer8c — Method

A linear least squares problem arising from chemical kinetics.

OptimizationProblems.PureJuMP.penalty2 — Method

A penalty problem by Gill, Murray and Pitfield in size 'n'

OptimizationProblems.PureJuMP.penalty3 — Method

A penalty problem by Gill, Murray and Pitfield in size 'n'

OptimizationProblems.PureJuMP.powellsg — Method

The extended Powell singular problem in size 'n'

OptimizationProblems.PureJuMP.power — Method

The Power problem by Oren.

OptimizationProblems.PureJuMP.quartc — Method

A simple quartic function.

OptimizationProblems.PureJuMP.rosenbrock — Method

Classic Rosenbrock problem

OptimizationProblems.PureJuMP.sbrybnd — Method

Broyden banded system of nonlinear equations in size 'n'

OptimizationProblems.PureJuMP.schmvett — Method

Another function with nontrivial groups and repetitious elements in size 'n'

OptimizationProblems.PureJuMP.scosine — Method

Another function with nontrivial groups and repetitious elements in size 'n'

OptimizationProblems.PureJuMP.sinquad — Method

Another function with nontrivial groups and repetitious elements in size 'n'

OptimizationProblems.PureJuMP.sparsine — Method

A sparse problem involving sine functions in size 'n'

OptimizationProblems.PureJuMP.sparsqur — Method

A sparse quartic problem in size 'n'

OptimizationProblems.PureJuMP.srosenbr — Method

The separable extension of Rosenbrock's function 'n'

OptimizationProblems.PureJuMP.tointgss — Method

Toint's Gaussian problem in size 'n'

OptimizationProblems.PureJuMP.tquartic — Method

A quartic function with nontrivial groups and repetitious elements in size 'n'

OptimizationProblems.PureJuMP.tridia — Method

Shanno's TRIDIA quadratic tridiagonal problem.

OptimizationProblems.PureJuMP.vardim — Method

Variable dimension problem.

OptimizationProblems.PureJuMP.woods — Method

The extended Woods problem n

OptimizationProblems.PureJuMP.zangwil3 — Method

Zangwil3 Model